Totally rework this entire library.
Okay, that may be a bit of an overstatement. What I realized was
that I wanted a utility function that had the same interface that
would properly handle staged files, unstaged changes, and full
files. I've created that interface but it required a bunch of
other changes to do it correctly.

I modified the names of addedlines, changedlines, and removedlines
so that they get grouped together when doing it alphabetically:
lines_added, lines_changed, lines_removed. That was totally unrelated.

The big thing is I created the get_diff_or_content(filenames) which
will give back a PatchSet, preferring staged changes, then unstaged
changes, then the entire content of the files requested.

This also means I'm committing to leaking unidiff.PatchSet instances
from this library. I had previously tried to avoid doing that. It's
just too useful at this point because some hooks are going to want
information on the lines added and removed and stuff. I could
reimplement that, but why?