Add "HEAD" to git diff-index command.
At least for me, in my recent tests, this command as originally written wasn't
working. I believe that my previous commit which changed "git diff" out for
"git diff-tree" was not correct and not fully tested because at the time I was
more concerned with adding support for the "--from"/"--to" arguments to
pre-commit.