Add more information to Unicode decode failures.
We recently hit an issue in a hook that uses this library which was incorrectly
performing a diff on a binary file. That's bad, but this library should handle
it gracefully. Before these changes we emitted a rather cryptic UnicodeError
which didn't catch the command that was run and therefore didn't expose the
file under analysis.

Additionally, we were attempting to erroneously decode the content as Unicode
a second time in our error message which just led to more decoding errors.
Instead we now just dump the bytestring which shows hex which is useful.